Add preliminary code implementation.
This code provides the bare-minimum functionality of a digital compass;
that is, it measures the heading from the qmc5883l magnetometer and
lights the appropriate leds on a 24-piece neopixel ring (i.e. in the
cardinal directions).

Future additions might include a debug mode for displaying magnetometer
readings via serial, adding a calibration mode instead of hardcoding a
single calibration, and potentially determining angle of declination via
something like gps.